refactor(ui): split api/client.ts — extract request core into requestCore.ts
client.ts was 565 lines mixing HTTP plumbing (case conversion, CSRF,
retry/backoff, request deduplication) with ~50 thin endpoint wrappers.
Moved the plumbing into requestCore.ts (303 lines); client.ts retains
only the typed endpoint shortcuts (~275 lines). Call sites still
import endpoints from '../api/client' as before. api/capture.ts now
imports the request primitive directly from requestCore.

Drops the last UI red-flag (>500 lines) file under budget.
